--- app/src/main/java/com/boostcamp/and03/data/api/NaverBookSearchApiConstants.kt
@@ -1,10 +1,0 @@
-package com.boostcamp.and03.data.api
-
-object NaverBookSearchApiConstants {
-    const val START_DEFAULT = 1
-    const val START_MAX = 1000
-    const val SIZE_DEFAULT = 10
-    const val SIZE_MAX = 100
-    const val SORT_SIMILARITY = "sim"
-    const val SORT_DATE = "date"
-}
